+++ server/file/models.py
@@ -12,15 +12,3 @@ class File(models.Model):
         File.objects.create(file=kwargs.get("files")).save()
 
 
-class Tag(models.Model):
-    tag = models.TextField("")
-
-    @classmethod
-    def init(cls, **kwargs):
-        Tag.objects.create(tag=kwargs.get("tag")).save()
-
-    @classmethod
-    def remove(cls, **kwargs):
-        Tag.objects.get(id=kwargs.get("id")).delete()
-
-
