+++ server/manage.py
@@ -23,14 +23,14 @@ if __name__ == '__main__':
 """
 python manage.py makemigrations budget
 python manage.py makemigrations classification
-python manage.py makemigrations documents
+python manage.py makemigrations file
 python manage.py makemigrations product
 python manage.py makemigrations accountOperations
 python manage.py makemigrations family
 python manage.py makemigrations
 python manage.py migrate budget
 python manage.py migrate classification
-python manage.py migrate documents
+python manage.py migrate file
 python manage.py migrate product
 python manage.py migrate accountOperations
 python manage.py migrate family
